services/backend/app/requirements/base.txt
fastapi
uvicorn
inflection
asyncpg
functools
tortoise-orm
fastapi-utils
psycopg2-binary
python-dotenv
aerich
